refactor: Standardize pixel value type to i32 for consistency
Changed all occurrences of 64-bit and 32-bit integer types to i32 to
ensure consistency across the codebase. This refactor eliminates type
discrepancies and improves maintainability.

+++ src/process/pixel_ops.c
@@ -3,7 +3,7 @@
 // ----------------------------------------------------------------
 int greyscale(Image *dst, const Image *src, size_t i, size_t j, void *ctx) {
     Pixel p = get_pixel(src, i, j);
-    int64_t grey = (int64_t)(0.299 * p.v[0] + 0.587 * p.v[1] + 0.114 * p.v[2]);
+    int32_t grey = (int32_t)(0.299 * p.v[0] + 0.587 * p.v[1] + 0.114 * p.v[2]);
     grey = CLAMP(grey, 0, 255);
 
     Pixel out = {0};
@@ -42,7 +42,7 @@ int threshold(Image *dst, const Image *src, size_t i, size_t j, void *ctx) {
     Pixel p = get_pixel(dst, i, j);
     Pixel out = {0};
 
-    int val = (p.v[0] >= threshold) ? 255 : 0;
+    int32_t val = (p.v[0] >= threshold) ? 255 : 0;
     out.v[0] = out.v[1] = out.v[2] = val;
     
     out.v[3] = p.v[3]; // preserve alpha if present
@@ -85,7 +85,7 @@ int contrast(Image *dst, const Image *src, size_t i, size_t j, void *ctx) {
 
     for (size_t c = 0; c < MIN(src->channels, 3); ++c) {
         float new_val = (float)(p.v[c] - 128) * factor + 128;
-        out.v[c] = CLAMP((int64_t)new_val, 0, 255);
+        out.v[c] = CLAMP((int32_t)new_val, 0, 255);
     }
 
     out.v[3] = p.v[3]; // preserve alpha if present
